+++ packages/react-native/src/alignment.ts
@@ -1,7 +1,7 @@
 import { type FlexAlignType, type FlexStyle } from 'react-native';
 
 export type Align = 'start' | 'center' | 'end' | 'stretch';
-export type Justify = 'start' | 'center' | 'end' | 'between';
+export type Justify = 'start' | 'center' | 'end' | 'between' | 'around';
 
 export const alignItems: Record<Align, FlexAlignType> = {
   center: 'center',
@@ -11,6 +11,7 @@ export const alignItems: Record<Align, FlexAlignType> = {
 };
 
 export const justifyContent: Record<Justify, FlexStyle['justifyContent']> = {
+  around: 'space-around',
   between: 'space-between',
   center: 'center',
   end: 'flex-end',